auth-service/requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
pydantic==1.10.7
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.7.0
python-multipart==0.0.6 